Move ShapeTestUtils and XShapeTestUtil to the test framework
